Moved additional components and additional components behavior to separate file to reduce clutter in manage course menu file
